Use HttpMethod
